Change the way benchmarking runs are resumed
Instead of at the root of the project, the pickle file is now stored in the directory of the interrupted benchmarking run.
Instead of providing the run id which corresponds to the file name of the pickle file, resuming is now controlled by providing the --resume-dir argument.
This behaviour is also closer to how it works in quark 2.0.